Refactor StateUtil class
